WIP: Implementing jammdb and refactoring
