+++ Cargo.toml
@@ -24,6 +24,6 @@ num-integer = "0.1.45"
 actix-cors = "0.6.4"
 serde-map-to-array = { version = "1.1.1", features = ["std"] }
 chrono = { version = "0.4.31", features = ["serde"] }
-persy = "1.4.5"
 actix-session = { version = "0.8.0", features = ["cookie-session"] }
 bcrypt = "0.15.0"
+jammdb = "0.11.0"
